Move jwtkey to an exp/support package
